Removed a ton of unneeded comments, changed NUI Message variable names.
Removed a whole lot of unneeded comment, so we can actually see the code underneath. Renamed the variables in the NUI Messages to be consistent with more recent projects. Will be making more changes very soon, just needed to clean up the clutter (comments) first. Tested in localhost, seems to work with no issues. Not really worth updating to, as nothing has changed.