Changed fire siren audio system
DO NOT BLINDLY UPDATE // THIS IS A DEVELOPMENT BUILD

Changed the audio system for the fire siren to match that of the pass alarm, in that instead of only being able to play one sound on the client at one time, now as many fire sirens as you want can be heard at the same time, meaning siren radiuses can now overlap without any issues. You can even stagger start them, if you want them to sound cooler. Tested on localhost with no issues. If you test this resource in your server, please report back to us! We have not thoroughly tested this update and would appreciate your feedback.