Added ability to add multiple fire siren sound files
DO NOT BLINDLY UPDATE // THIS IS A DEVELOPMENT BUILD

Using the new audio system from the previous commit, I've added the ability for multiple different fire sirens to be used, and set per-station in the config. When adding new siren files, try to keep them under 1mb, or you might run into streaming issues. Tested on localhost with no issues. If you test this resource in your server, please report back to us! We have not thoroughly tested this update and would appreciate your feedback.